requirements.txt
mysql-connector-python == 8.0.27
pysimplegui == 4.60.4